Fix feature support
Add a no-op code handle for subscription creation when the crate is compiled
without the mint feature
